+++ feature/events/src/main/java/com/arshapshap/events/domain/EventsInteractor.kt
@@ -1,6 +1,6 @@
 package com.arshapshap.events.domain
 
-import com.arshapshap.events.domain.interfaces.EventsRepository
+import com.arshapshap.events.domain.repositories.EventsRepository
 import com.arshapshap.events.domain.models.Event
 import javax.inject.Inject
 
@@ -12,11 +12,11 @@ class EventsInteractor @Inject constructor(
         return repository.getEvents()
     }
 
-    internal suspend fun getEventById(id: Int): Event? {
+    internal suspend fun getEventById(id: Long): Event? {
         return repository.getEventById(id)
     }
 
-    internal suspend fun deleteEventById(id: Int) {
+    internal suspend fun deleteEventById(id: Long) {
         repository.deleteEventById(id)
     }
 }
