+++ feature/events/src/main/java/com/arshapshap/events/presentation/screens/EventsFeatureRouter.kt
@@ -2,7 +2,7 @@ package com.arshapshap.events.presentation.screens
 
 interface EventsFeatureRouter {
 
-    fun openEvent(id: Int)
+    fun openEvent(id: Long)
 
     fun closeCurrentFragment()
 }
